clean user controller
