+++ src/routes.ts
@@ -7,7 +7,8 @@ import UserController from './app/controllers/UserController';
 
 const router = Router();
 
-router.get('/users', UserController.index)
+router.get('/users', UserController.index);
+router.get('/users/:id', UserController.show);
 router.post('/users', UserController.store);
 
 router.post('/auth', AuthController.authenticate);
